requirements.txt
attrs==25.4.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
marko==2.1.2
MarkupSafe==2.1.5
PyYAML==6.0.2
referencing==0.37.0
rpds-py==0.30.0
typing_extensions==4.15.0
